operator: move CreateNamespace to pkg/util
