+++ operator/cmd/mesh/shared.go
@@ -203,11 +203,6 @@ func getCRAndNamespaceFromFile(filePath string, l clog.Logger) (customResource s
 	return
 }
 
-// createNamespace creates a namespace using the given k8s interface.
-func createNamespace(cs kubernetes.Interface, namespace string, network string, dryRun bool) error {
-	return helmreconciler.CreateNamespace(cs, namespace, network, dryRun)
-}
-
 // saveIOPToCluster saves the state in an IOP CR in the cluster.
 func saveIOPToCluster(reconciler *helmreconciler.HelmReconciler, iop string) error {
 	obj, err := object.ParseYAMLToK8sObject([]byte(iop))
